Extract Creator Kit browse/read clusters out of mcp-server.ts
Continues Wave C (batch 2). Moves the 7-tool kit cluster
(get_kit/get_kit_api/list_kit_files/search_kit_files/read_kit_file/
read_kit_files/read_kit_file_fragment) into two new files —
mcp-kit-tools.ts (orientation: fetch/API digest/list/search) and
mcp-kit-file-tools.ts (read: single/batch/fragment) — split to stay
under the 500-line new-file cap. Promotes KIT_ENGINE_REF_PROP and
MCP_VISIBLE_TOOLS into the shared mcp-tool-support.ts (both had
external call sites in mcp-server.ts beyond this cluster).
