+++ apps/api/src/agent-surface/mcp-server.ts
@@ -7,11 +7,17 @@ import {
   toolErr,
   BEHAVIOURAL_CONTRACT,
   SESSION_KEY_PROP,
+  KIT_ENGINE_REF_PROP,
+  MCP_VISIBLE_TOOLS,
   type ToolResult,
   type ToolHandler,
   type ToolContext,
 } from './mcp-tool-support.js';
 import { createExampleTools } from './mcp-example-tools.js';
+import { createKitTools } from './mcp-kit-tools.js';
+import { createKitFileTools } from './mcp-kit-file-tools.js';
+
+export { MCP_VISIBLE_TOOLS };
 import { looksLikeCreatorAgentKey } from './agent-creator-key.js';
 import { canonicalAppBaseUrl } from '../platform/canonical-app-url.js';
 import { DEFAULT_TRANSCRIPT_WINDOW_ENTRIES, MAX_TRANSCRIPT_WINDOW_ENTRIES } from '../delivery/build-transcript.js';
@@ -143,46 +149,6 @@ const FINISHED_REASON = STALE_AGENT_TOKEN_REASON;
 
 const SESSION_HEADER = 'mcp-session-id';
 
-export const MCP_VISIBLE_TOOLS = new Set([
-  'create_game',
-  'start',
-  'open_round',
-  'continue_draft',
-  'get_brief',
-  'get_seed',
-  'regenerate_seed',
-  'get_sources',
-  'get_kit',
-  // get_kit_api is the orientation path; browse tools are the depth path.
-  'get_kit_api',
-  'list_kit_files',
-  'search_kit_files',
-  'read_kit_file',
-  'read_kit_files',
-  'read_kit_file_fragment',
-  'knowledge_query',
-  'report_progress',
-  'screenshot_upload_url',
-  'stage_upload_url',
-  'stage_source_file',
-  'patch_source_file',
-  'list_staged_sources',
-  'clear_staged_sources',
-  'delete_source_file',
-  'submit_sources',
-  'end',
-  'show_round',
-  'show_media',
-  'get_round_status',
-  'get_gate_verdict',
-  'get_gate_media',
-  'get_round_media',
-  'get_reference_images',
-  'read_inbox',
-  'ack_inbox',
-  'get_transcript',
-]);
-
 // Callable for REST clients, never advertised, never named to a model.
 export const MCP_UNADVERTISED_TOOLS: readonly string[] = Object.freeze([
   'open_proposal_round',
@@ -193,13 +159,6 @@ export const MCP_UNADVERTISED_TOOLS: readonly string[] = Object.freeze([
   'list_example_files',
   'read_example_file',
 ]);
-
-function withAdvertisedBrowseTools<T extends { browse?: Record<string, string> }>(body: T): T {
-  if (!body.browse) return body;
-  const advertised = Object.entries(body.browse).filter(([, tool]) => MCP_VISIBLE_TOOLS.has(tool));
-  const { browse: _browse, ...rest } = body;
-  return (advertised.length ? { ...rest, browse: Object.fromEntries(advertised) } : rest) as T;
-}
 
 /** Aggressive ceiling on unauthenticated / invalid `start` attempts per IP. */
 const MAX_INVALID_STARTS_PER_WINDOW = 20;
@@ -724,13 +683,6 @@ function startToolResult(structured: { sessionKey: string } & Record<string, unk
  * argument against it is a type error rather than a check.
  */
 const BUILD_STEP_NAMES: ReadonlySet<string> = new Set<string>(BUILD_STEPS);
-
-/** Pin kit browse/read calls to the engineRef get_kit returned (N/N−1 window). */
-const KIT_ENGINE_REF_PROP = {
-  type: 'string' as const,
-  description:
-    'Creator Kit engineRef from get_kit. Pass on every browse/read call so a mid-round registry bump cannot mix kit revisions.',
-};
 
 export async function registerMcpServerRoutes(app: FastifyInstance, options: McpServerOptions = {}): Promise<void> {
   const store = options.store;
@@ -2694,473 +2646,8 @@ export async function registerMcpServerRoutes(app: FastifyInstance, options: Mcp
       },
     },
 
-    get_kit: {
-      annotations: { title: 'Fetch the Creator Kit', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          engineRef: { type: 'string' },
-          kitUrl: { type: 'string' },
-          sha256: { type: 'string' },
-          unpack: { type: 'string' },
-          entry: { type: 'string' },
-          kitEngineChanged: { type: 'boolean' },
-          browse: {
-            type: 'object',
-            properties: {
-              list: { type: 'string' },
-              search: { type: 'string' },
-              read: { type: 'string' },
-              readMany: { type: 'string' },
-              fragment: { type: 'string' },
-            },
-          },
-        },
-        required: ['engineRef', 'kitUrl', 'sha256', 'unpack', 'entry'],
-      },
-      description:
-        'Fetch Creator Kit metadata: engineRef (required for submit_sources), sha256, entry, ' +
-        'and optional kitUrl/unpack for agents with shell egress. ' +
-        'engineRef is pinned for the round: repeat calls return the same engine even if the ' +
-        'registry moves. kitEngineChanged:true means the pin was replaced — after a kit_outdated ' +
-        'verdict, or because the pinned kit is no longer retained — so rebuild against the ' +
-        'engine in this reply. ' +
-        'This platform and its Creator Kit are not on the public web — an unanswered question ' +
-        'about what it can build (multiplayer, persistent worlds, party games, …) is answered by ' +
-        'get_kit_api or browse, never by web search. ' +
-        'For the API itself: get_kit_api (with this engineRef) for a prompt-ready orientation in ' +
-        'one call — it flags what it had to cut (by name when a whole declaration is dropped, by ' +
-        'count when a kept one is trimmed member-wise), so a missing signature is never silent; ' +
-        "use the browse tools named in this reply's browse block (list/search/read) for those or " +
-        'any other specific kit file. ' +
-        'With shell egress, kitUrl/unpack lets you unpack the kit locally and read SKILL.md directly. ' +
-        'entry=gamedevpl-creator-kit/SKILL.md (tarball roots at gamedevpl-creator-kit/; ' +
-        'do not assume a `cd` persists across tool calls). ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: { sessionKey: SESSION_KEY_PROP },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const res = await injectChannel(ctx.request, 'GET', AGENT_CHANNEL_ROUTES.KIT, auth.channelToken);
-        const body = res.json() as { error?: string; message?: string; browse?: Record<string, string> };
-        if (res.statusCode !== 200) {
-          return toolErr(body.message ?? body.error ?? `kit failed (${res.statusCode})`, body);
-        }
-        return toolOk(withAdvertisedBrowseTools(body));
-      },
-    },
-
-    get_kit_api: {
-      annotations: { title: "Fetch the Creator Kit's API reference", ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          engineRef: { type: 'string' },
-          digest: { type: 'string' },
-        },
-        required: ['engineRef', 'digest'],
-      },
-      description:
-        "The Creator Kit's prompt-ready orientation in one call: what engine modules exist " +
-        '(party for same-screen multiplayer, zone for a real-time server-arbitrated world, ' +
-        'commons and presence for persistent/shared state, and the rest — this is the answer ' +
-        'to "can this platform build X", not a web search), plus as much of the core API ' +
-        'signatures, audio catalog, and exemplar game as fit in one tool result. ' +
-        'The response is sized to a safe single-call limit, not to the whole API — for a real ' +
-        'kit this routinely omits content: whole declarations dropped are named in an "Omitted ' +
-        'for length" note, and a declaration too large to fit whole is trimmed member-wise with ' +
-        'only a count of what was cut, not names. Treat both as normal, not an error. ' +
-        'Call this once near ' +
-        'the start of a round, before scaffolding, rather than repeatedly — its content only ' +
-        'changes when engineRef does. Pass engineRef from get_kit so a mid-round registry bump ' +
-        "cannot mix kit revisions. Falls back to the registry's current engine when engineRef is " +
-        'omitted, but that risks reading a different kit than the round is pinned to. ' +
-        'Gives that orientation without unpacking the whole kit into context; the browse tools ' +
-        '(list_kit_files / search_kit_files / read_kit_file) cover anything this digest omitted, ' +
-        'summarized, or named in its omission note. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: { sessionKey: SESSION_KEY_PROP, engineRef: KIT_ENGINE_REF_PROP },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const params = new URLSearchParams();
-        if (typeof args.engineRef === 'string' && args.engineRef.trim()) {
-          params.set('engineRef', args.engineRef.trim());
-        }
-        const qs = params.toString();
-        const res = await injectChannel(
-          ctx.request,
-          'GET',
-          `${AGENT_CHANNEL_ROUTES.KIT_API}${qs ? `?${qs}` : ''}`,
-          auth.channelToken,
-        );
-        const body = res.json() as { error?: string; message?: string };
-        if (res.statusCode !== 200) {
-          return toolErr(body.message ?? body.error ?? `get_kit_api failed (${res.statusCode})`, body);
-        }
-        return toolOk(body);
-      },
-    },
-
-    list_kit_files: {
-      annotations: { title: 'List Creator Kit files', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          engineRef: { type: 'string' },
-          entry: { type: 'string' },
-          files: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                path: { type: 'string' },
-                bytes: { type: 'number' },
-                kind: { type: 'string', enum: ['text', 'binary'] },
-              },
-              required: ['path', 'bytes', 'kind'],
-            },
-          },
-          total: { type: 'number' },
-          truncated: { type: 'boolean' },
-        },
-        required: ['engineRef', 'entry', 'files', 'total', 'truncated'],
-      },
-      description:
-        'List paths inside a pinned Creator Kit (size + text/binary kind). ' +
-        'Pass engineRef from get_kit. Optional prefix (e.g. shared/modules) or simple glob (*). ' +
-        'Paginate with limit/offset. Start from get_kit.entry via read_kit_file. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          engineRef: KIT_ENGINE_REF_PROP,
-          prefix: { type: 'string', description: 'Path prefix under the kit root (or full gamedevpl-creator-kit/…).' },
-          glob: { type: 'string', description: 'Simple glob with * wildcards (e.g. **/*.md or shared/modules/*.ts).' },
-          limit: { type: 'integer', description: 'Max paths to return (default 200, max 500).' },
-          offset: { type: 'integer', description: 'Skip this many matching paths.' },
-        },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const params = new URLSearchParams();
-        if (typeof args.engineRef === 'string' && args.engineRef.trim()) {
-          params.set('engineRef', args.engineRef.trim());
-        }
-        if (typeof args.prefix === 'string' && args.prefix.trim()) params.set('prefix', args.prefix.trim());
-        if (typeof args.glob === 'string' && args.glob.trim()) params.set('glob', args.glob.trim());
-        if (typeof args.limit === 'number' && Number.isFinite(args.limit)) params.set('limit', String(args.limit));
-        if (typeof args.offset === 'number' && Number.isFinite(args.offset)) params.set('offset', String(args.offset));
-        const qs = params.toString();
-        const res = await injectChannel(
-          ctx.request,
-          'GET',
-          `${AGENT_CHANNEL_ROUTES.KIT_FILES}${qs ? `?${qs}` : ''}`,
-          auth.channelToken,
-        );
-        const body = res.json() as { error?: string; message?: string };
-        if (res.statusCode !== 200) {
-          return toolErr(body.message ?? body.error ?? `list_kit_files failed (${res.statusCode})`, body);
-        }
-        return toolOk(body);
-      },
-    },
-
-    search_kit_files: {
-      annotations: { title: 'Search Creator Kit files', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          engineRef: { type: 'string' },
-          query: { type: 'string' },
-          matches: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                path: { type: 'string' },
-                line: { type: 'number' },
-                text: { type: 'string' },
-              },
-              required: ['path', 'line', 'text'],
-            },
-          },
-          truncated: { type: 'boolean' },
-          filesScanned: { type: 'number' },
-        },
-        required: ['engineRef', 'query', 'matches', 'truncated', 'filesScanned'],
-      },
-      description:
-        'Search text files in a pinned Creator Kit for a substring (case-insensitive). ' +
-        'Pass engineRef from get_kit. Returns path + line + snippet; capped match count. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          engineRef: KIT_ENGINE_REF_PROP,
-          query: { type: 'string', description: 'Substring to find (2–120 chars).' },
-          prefix: { type: 'string', description: 'Optional path prefix to narrow the search.' },
-          limit: { type: 'integer', description: 'Max matches (default/max 40).' },
-        },
-        required: ['query'],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const query = typeof args.query === 'string' ? args.query.trim() : '';
-        if (!query) return toolErr('query is required');
-        const params = new URLSearchParams({ q: query });
-        if (typeof args.engineRef === 'string' && args.engineRef.trim()) {
-          params.set('engineRef', args.engineRef.trim());
-        }
-        if (typeof args.prefix === 'string' && args.prefix.trim()) params.set('prefix', args.prefix.trim());
-        if (typeof args.limit === 'number' && Number.isFinite(args.limit)) params.set('limit', String(args.limit));
-        const res = await injectChannel(
-          ctx.request,
-          'GET',
-          `${AGENT_CHANNEL_ROUTES.KIT_SEARCH}?${params.toString()}`,
-          auth.channelToken,
-        );
-        const body = res.json() as { error?: string; message?: string };
-        if (res.statusCode !== 200) {
-          return toolErr(body.message ?? body.error ?? `search_kit_files failed (${res.statusCode})`, body);
-        }
-        return toolOk(body);
-      },
-    },
-
-    read_kit_file: {
-      annotations: { title: 'Read one Creator Kit file', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          engineRef: { type: 'string' },
-          path: { type: 'string' },
-          bytes: { type: 'number' },
-          kind: { type: 'string', enum: ['text', 'binary'] },
-          encoding: { type: 'string', enum: ['utf8', 'base64'] },
-          content: { type: 'string' },
-        },
-        required: ['engineRef', 'path', 'bytes', 'kind', 'encoding', 'content'],
-      },
-      description:
-        'Read one small Creator Kit file (≤48 KiB). Use read_kit_files when fetching several known paths. ' +
-        'Pass engineRef from get_kit. Larger files return kit_file_too_large — use read_kit_file_fragment. ' +
-        'Binary files need encoding=base64. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          engineRef: KIT_ENGINE_REF_PROP,
-          path: { type: 'string', description: 'Kit file path (e.g. SKILL.md or gamedevpl-creator-kit/SKILL.md).' },
-          encoding: {
-            type: 'string',
-            enum: ['utf8', 'base64'],
-            description: 'utf8 for text (default); base64 required for binary.',
-          },
-        },
-        required: ['path'],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const path = typeof args.path === 'string' ? args.path.trim() : '';
-        if (!path) return toolErr('path is required');
-        const params = new URLSearchParams({ path });
-        if (typeof args.engineRef === 'string' && args.engineRef.trim()) {
-          params.set('engineRef', args.engineRef.trim());
-        }
-        if (args.encoding === 'base64' || args.encoding === 'utf8') params.set('encoding', args.encoding);
-        const res = await injectChannel(
-          ctx.request,
-          'GET',
-          `${AGENT_CHANNEL_ROUTES.KIT_FILE}?${params.toString()}`,
-          auth.channelToken,
-        );
-        const body = res.json() as { error?: string; message?: string };
-        if (res.statusCode !== 200) {
-          return toolErr(body.message ?? body.error ?? `read_kit_file failed (${res.statusCode})`, body);
-        }
-        return toolOk(body);
-      },
-    },
-
-    read_kit_files: {
-      annotations: { title: 'Read several Creator Kit files', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          engineRef: { type: 'string' },
-          files: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                ok: { type: 'boolean' },
-                path: { type: 'string' },
-                bytes: { type: 'number' },
-                kind: { type: 'string', enum: ['text', 'binary'] },
-                encoding: { type: 'string', enum: ['utf8', 'base64'] },
-                content: { type: 'string' },
-                error: { type: 'string' },
-                message: { type: 'string' },
-              },
-              required: ['ok', 'path'],
-            },
-          },
-          totalBytes: { type: 'number' },
-          maxBytes: { type: 'number' },
-          maxFiles: { type: 'number' },
-          truncated: { type: 'boolean' },
-        },
-        required: ['engineRef', 'files', 'totalBytes', 'maxBytes', 'maxFiles', 'truncated'],
-      },
-      description:
-        'Read up to 12 small Creator Kit files in one call (≤128 KiB aggregate), staying within per-turn ' +
-        'tool-call limits. Pass engineRef from get_kit. ' +
-        'Per-path failures stay in files[]; oversized files need read_kit_file_fragment. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          engineRef: KIT_ENGINE_REF_PROP,
-          paths: {
-            type: 'array',
-            items: { type: 'string' },
-            description: 'Kit file paths (1–12), e.g. ["SKILL.md", "starters/block-cascade/game.ts"].',
-          },
-          encoding: {
-            type: 'string',
-            enum: ['utf8', 'base64'],
-            description: 'Optional override for every file; default is utf8 for text and base64 for binary.',
-          },
-        },
-        required: ['paths'],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        if (!Array.isArray(args.paths)) return toolErr('paths must be an array of strings');
-        const paths = args.paths.filter((path): path is string => typeof path === 'string');
-        if (paths.length === 0) return toolErr('paths must be a non-empty array');
-        const body: Record<string, unknown> = { paths };
-        if (typeof args.engineRef === 'string' && args.engineRef.trim()) {
-          body.engineRef = args.engineRef.trim();
-        }
-        if (args.encoding === 'base64' || args.encoding === 'utf8') body.encoding = args.encoding;
-        const res = await injectChannel(
-          ctx.request,
-          'POST',
-          AGENT_CHANNEL_ROUTES.KIT_FILES_READ,
-          auth.channelToken,
-          body,
-        );
-        const payload = res.json() as { error?: string; message?: string };
-        if (res.statusCode !== 200) {
-          return toolErr(payload.message ?? payload.error ?? `read_kit_files failed (${res.statusCode})`, payload);
-        }
-        return toolOk(payload);
-      },
-    },
-
-    read_kit_file_fragment: {
-      annotations: { title: 'Read a Creator Kit file fragment', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          engineRef: { type: 'string' },
-          path: { type: 'string' },
-          kind: { type: 'string', enum: ['text', 'binary'] },
-          unit: { type: 'string', enum: ['bytes', 'lines'] },
-          offset: { type: 'number' },
-          limit: { type: 'number' },
-          totalBytes: { type: 'number' },
-          totalLines: { type: ['number', 'null'] },
-          encoding: { type: 'string', enum: ['utf8', 'base64'] },
-          content: { type: 'string' },
-          eof: { type: 'boolean' },
-          nextOffset: { type: ['number', 'null'] },
-        },
-        required: [
-          'engineRef',
-          'path',
-          'kind',
-          'unit',
-          'offset',
-          'limit',
-          'totalBytes',
-          'totalLines',
-          'encoding',
-          'content',
-          'eof',
-          'nextOffset',
-        ],
-      },
-      description:
-        'Read a window of one Creator Kit file by lines (default) or bytes (always base64). ' +
-        'Pass engineRef from get_kit. Use nextOffset for pagination. Overlong line windows error — switch to unit=bytes. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          engineRef: KIT_ENGINE_REF_PROP,
-          path: { type: 'string', description: 'Kit file path.' },
-          offset: { type: 'integer', description: '0-based start line or byte (use nextOffset from the prior reply).' },
-          limit: { type: 'integer', description: 'Max lines (≤200) or bytes (≤32 KiB).' },
-          unit: {
-            type: 'string',
-            enum: ['lines', 'bytes'],
-            description: 'Default lines; bytes required for binary and always returns base64.',
-          },
-          encoding: {
-            type: 'string',
-            enum: ['utf8', 'base64'],
-            description: 'utf8 for lines; base64 required for unit=bytes.',
-          },
-        },
-        required: ['path'],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const path = typeof args.path === 'string' ? args.path.trim() : '';
-        if (!path) return toolErr('path is required');
-        const params = new URLSearchParams({ path });
-        if (typeof args.engineRef === 'string' && args.engineRef.trim()) {
-          params.set('engineRef', args.engineRef.trim());
-        }
-        if (typeof args.offset === 'number' && Number.isFinite(args.offset)) params.set('offset', String(args.offset));
-        if (typeof args.limit === 'number' && Number.isFinite(args.limit)) params.set('limit', String(args.limit));
-        if (args.unit === 'lines' || args.unit === 'bytes') params.set('unit', args.unit);
-        if (args.encoding === 'base64' || args.encoding === 'utf8') params.set('encoding', args.encoding);
-        const res = await injectChannel(
-          ctx.request,
-          'GET',
-          `${AGENT_CHANNEL_ROUTES.KIT_FILE_FRAGMENT}?${params.toString()}`,
-          auth.channelToken,
-        );
-        const body = res.json() as { error?: string; message?: string };
-        if (res.statusCode !== 200) {
-          return toolErr(body.message ?? body.error ?? `read_kit_file_fragment failed (${res.statusCode})`, body);
-        }
-        return toolOk(body);
-      },
-    },
+    ...createKitTools({ resolveAuth, injectChannel }),
+    ...createKitFileTools({ resolveAuth, injectChannel }),
 
     knowledge_query: {
       annotations: { title: 'Query GameKit/EditorKit/example-game knowledge', ...READS },
