+++ eslint-rules/module-boundary-map.mjs
@@ -177,6 +177,8 @@ const FILE_BUCKET = {
   'mcp-server': 'agent-surface',
   'mcp-tool-support': 'agent-surface',
   'mcp-example-tools': 'agent-surface',
+  'mcp-kit-tools': 'agent-surface',
+  'mcp-kit-file-tools': 'agent-surface',
   'mcp-server-discovery': 'agent-surface',
   'mcp-ui': 'agent-surface',
   'mcp-oauth-metadata': 'agent-surface',
